Move some python files from the top-level source directory
